+++ src/driver_PORT.h
@@ -1,6 +1,9 @@
 #pragma once
 
+#include <stddef.h>
 #include <stdint.h>
+
+#include "emonTH_saml.h"
 
 /* Types */
 typedef struct Pin_ {
@@ -20,35 +23,35 @@ typedef enum PINDRV_ { PIN_DRV_CLR, PIN_DRV_SET, PIN_DRV_TGL } PINDRV_t;
  *  @param [in] cfg : configuration option
  *  @param [in] cs  : clear or set configuration
  */
-void portPinCfg(unsigned int pin, unsigned int cfg, PINCFG_t cs);
+void portPinCfg(size_t pin, unsigned int cfg, PINCFG_t cs);
 
 /*! @brief Sets a pin as input or output
  *  @param [in] pin : PIN number
  *  @param [in] mode: PIN_DIR_IN for input, PIN_DIR_OUT for output
  */
-void portPinDir(unsigned int pin, PINDIR_t mode);
+void portPinDir(size_t pin, PINDIR_t mode);
 
 /*! @brief Sets the pin driver value
  *  @param [in] pin : Pin number
  *  @param [in] drv : Clear, set, or toggle pin
  */
-void portPinDrv(unsigned int pin, PINDRV_t drv);
+void portPinDrv(size_t pin, PINDRV_t drv);
 
 /*! @brief Sets the mux for pin alternate function
  *  @param [in] pin : Pin number
  *  @param [in] mux : Mux mode
  */
-void portPinMux(unsigned int pin, unsigned int mux);
+void portPinMux(size_t pin, unsigned int mux);
 
 /*! @brief Clear the mux for pin alternate function
  *  @param [in] pin : Pin number
  */
-void portPinMuxClear(unsigned int pin);
+void portPinMuxClear(size_t pin);
 
 /*! @brief Returns the pin value
  *  @param [in] pin : Pin number
  */
-unsigned int portPinValue(unsigned int pin);
+unsigned int portPinValue(size_t pin);
 
 /*! @brief   Configure the ports.
  *           Ports for peripherals are configured in their setup functions
